requirements.txt
setuptools
wheel
pydantic
pyyaml
fastapi
uvicorn[standard]
starlette-exporter
starlette
pytest
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
azure-identity
azure-storage-blob
prometheus-client
pytest
pytest-asyncio
google-cloud-storage>=2.15.0
google-auth>=2.23.0
